+++ src/main/java/sk/frisbee/jdbc/SearchDaoImpl.java
@@ -58,8 +58,8 @@ public class SearchDaoImpl implements SearchDao {
 		@Override
 		public Player mapRow(ResultSet rs, int rowNum) throws SQLException {
 			Player player = new Player();
-			player.setPlayerId(rs.getLong("id_hrac"));
-			player.setUserId(rs.getLong("id_user"));
+			player.setPlayerId(rs.getInt("id_hrac"));
+			player.setUserId(rs.getInt("id_user"));
 			player.setFirstName(rs.getString("meno"));
 			player.setLastName(rs.getString("priezvisko"));
 			player.setDisciplines(rs.getString("discipliny"));
